Rename and move INonDet to program.NondeterministicExpression
