+++ dartagnan/src/main/java/com/dat3m/dartagnan/expression/processing/ExpressionVisitor.java
@@ -1,6 +1,7 @@
 package com.dat3m.dartagnan.expression.processing;
 
 import com.dat3m.dartagnan.expression.*;
+import com.dat3m.dartagnan.program.NondeterministicExpression;
 import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.expression.*;
 import com.dat3m.dartagnan.program.memory.Location;
@@ -14,7 +15,7 @@ public interface ExpressionVisitor<T> {
     default T visit(IExprBin iBin) { return null; }
     default T visit(IExprUn iUn) { return null; }
     default T visit(ConditionalExpression ifExpr) { return null; }
-    default T visit(INonDet iNonDet) { return null; }
+    default T visit(NondeterministicExpression iNonDet) { return null; }
     default T visit(Register reg) { return null; }
     default T visit(MemoryObject address) { return null; }
     default T visit(Location location) { return null; }
